Another attempt at the example
