+++ example/App.tsx
@@ -1,118 +1,6 @@
-import { ExpoWebGLRenderingContext, GLView } from 'expo-gl';
-import { loadAsync, Renderer, TextureLoader } from 'expo-three';
 import * as React from 'react';
-import {
-  AmbientLight,
-  BoxGeometry,
-  Fog,
-  GridHelper,
-  Mesh,
-  MeshStandardMaterial,
-  PerspectiveCamera,
-  PointLight,
-  Scene,
-  SpotLight,
-} from 'three';
+import App from './src/App';
 
-export default function App() {
-  const timeoutRef = React.useRef<number>();
-
-  React.useEffect(() => {
-    // Clear the animation loop when the component unmounts
-    return () => clearTimeout(timeoutRef.current);
-  }, []);
-
-  const onContextCreate = async (gl: ExpoWebGLRenderingContext) => {
-    // removes the warning EXGL: gl.pixelStorei() doesn't support this parameter yet!
-    const pixelStorei = gl.pixelStorei.bind(gl);
-    gl.pixelStorei = function (...args) {
-      const [parameter] = args;
-      switch (parameter) {
-        case gl.UNPACK_FLIP_Y_WEBGL:
-          return pixelStorei(...args);
-      }
-    };
-
-    const { drawingBufferWidth: width, drawingBufferHeight: height } = gl;
-    const sceneColor = 0x6ad6f0;
-
-    // Create a WebGLRenderer without a DOM element
-    const renderer = new Renderer({ gl });
-    renderer.setSize(width, height);
-    renderer.setClearColor(sceneColor);
-
-    const camera = new PerspectiveCamera(70, width / height, 0.01, 1000);
-    camera.position.set(2, 5, 5);
-
-    const scene = new Scene();
-    scene.fog = new Fog(sceneColor, 1, 10000);
-    scene.add(new GridHelper(10, 10));
-
-    const ambientLight = new AmbientLight(0x101010);
-    scene.add(ambientLight);
-
-    const pointLight = new PointLight(0xffffff, 2, 1000, 1);
-    pointLight.position.set(0, 200, 200);
-    scene.add(pointLight);
-
-    const spotLight = new SpotLight(0xffffff, 0.5);
-    spotLight.position.set(0, 500, 100);
-    spotLight.lookAt(scene.position);
-    scene.add(spotLight);
-
-    // Load and add a texture
-    const cube = new IconMesh();
-    scene.add(cube);
-
-    // Load and add an obj model
-    const model: Record<string, string> = {
-      '3d.obj':
-        'https://raw.githubusercontent.com/mrdoob/three.js/dev/examples/models/obj/walt/WaltHead.obj',
-      '3d.mtl':
-        'https://raw.githubusercontent.com/mrdoob/three.js/dev/examples/models/obj/walt/WaltHead.mtl',
-    };
-
-    const object = await loadAsync(
-      [model['3d.obj'], model['3d.mtl']],
-      // @ts-ignore
-      null,
-      (name) => model[name]
-    );
-
-    object.position.y += 2;
-    object.position.z -= 2;
-    object.scale.set(0.02, 0.02, 0.02);
-
-    scene.add(object);
-
-    camera.lookAt(cube.position);
-
-    function update() {
-      cube.rotation.y += 0.05;
-      cube.rotation.x += 0.025;
-    }
-
-    // Setup an animation loop
-    const render = () => {
-      timeoutRef.current = requestAnimationFrame(render);
-      update();
-      renderer.render(scene, camera);
-      gl.endFrameEXP();
-    };
-    render();
-  };
-
-  return <GLView style={{ flex: 1 }} onContextCreate={onContextCreate} />;
-}
-
-class IconMesh extends Mesh {
-  constructor() {
-    super(
-      new BoxGeometry(1.0, 1.0, 1.0),
-      new MeshStandardMaterial({
-        map: new TextureLoader().load(require('./assets/icon.png')),
-        // color: 0xff0000
-      })
-    );
-  }
+export default function RootApp() {
+  return <App />;
 }
