+++ example/package.json
@@ -1,48 +1,35 @@
 {
-  "name": "ExpoThreeExample",
-  "version": "0.0.1",
-  "private": true,
+  "name": "expo-three-example",
+  "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
-    "compile": "tsc --noEmit -p . --pretty",
-    "start": "npx expo start",
-    "android": "npx expo run:android",
-    "ios": "npx expo run:ios",
-    "web": "npx expo start --web",
-    "bundle:web": "npx expo export --platform web",
-    "serve:web": "npx server dist",
-    "prebuild:clean": "npx expo prebuild --clean"
+    "start": "expo start",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
+    "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/metro-runtime": "~3.1.3",
-    "expo": "^50.0.0",
+    "expo": "~50.0.11",
+    "expo-asset": "~9.0.2",
+    "expo-asset-utils": "^3.0.0",
     "expo-build-properties": "~0.11.1",
-    "expo-constants": "~15.4.5",
-    "expo-dev-client": "~3.3.6",
     "expo-gl": "~13.6.0",
-    "expo-linking": "~6.2.2",
-    "expo-status-bar": "~1.11.1",
-    "expo-three": "^7.0.0",
+    "expo-module-scripts": "^3.4.1",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "0.73.4",
-    "react-native-web": "~0.19.6",
-    "three": "^0.145.0"
+    "react-native": "0.73.5",
+    "three": "^0.154.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
-    "@babel/preset-env": "^7.20.0",
-    "@babel/runtime": "^7.20.0",
-    "@types/react": "~18.2.14",
-    "@types/three": "^0.162.0",
-    "@typescript-eslint/eslint-plugin": "^5.59.0",
-    "@typescript-eslint/parser": "^5.59.0",
-    "babel-plugin-module-resolver": "^5.0.0",
-    "prettier": "2.8.8",
+    "@types/react": "~18.2.45",
+    "@types/three": "^0.154.0",
     "ts-node": "^10.9.2",
-    "typescript": "5.3.2"
+    "typescript": "^5.1.3"
   },
-  "engines": {
-    "node": ">=18"
+  "private": true,
+  "expo": {
+    "autolinking": {
+      "nativeModulesDir": ".."
+    }
   }
 }
