+++ example/src/icon.png
